services/embedding-engine/requirements.txt
redis==5.0.1
sentence-transformers==2.2.2
weaviate-client==3.25.3
torch==2.1.1
transformers==4.36.0
numpy==1.24.3
python-dotenv==1.0.0
structlog==23.2.0
